+++ adapters/logrusadapter/logger_test.go
@@ -40,7 +40,3 @@ func newTestSuite() *loggertesting.LoggerTestSuite {
 func TestLogger_Levels(t *testing.T) {
 	newTestSuite().TestLevels(t)
 }
-
-func TestLogger_Levelsln(t *testing.T) {
-	newTestSuite().TestLevelsln(t)
-}