+++ adapters/simplelogadapter/logger_test.go
@@ -35,7 +35,3 @@ func newTestSuite() *loggertesting.LoggerTestSuite {
 func TestLogger_Levels(t *testing.T) {
 	newTestSuite().TestLevels(t)
 }
-
-func TestLogger_Levelsln(t *testing.T) {
-	newTestSuite().TestLevelsln(t)
-}